Modify SmallDataIO
* Move constructors to cpp file.
* The step number in the filename when writing in NEW mode is now padded
(with zeros) to be a specifiable number of digits (defaulted to 6).
* Add data and coords epsilons to SmallDataIO: the epsilons are no
longer hardcoded to 1.0e-8 and are instead related to the data and coord
precisions. There are also getters and static default getters.
* Add a new public member function which returns the full filename for
files written in NEW mode at a specific time (for given dt).
* Add a random integer to the temporary filename in
remove_duplicate_time_data.
* Move some includes from the header to the cpp file.
* Remove some member variables only used in the constructor.

Also sync with CTC repository.